fix(server): upgrade packages to resolve vulnerabilities, update socket.io version
This commit addresses the issue of detected vulnerabilities in our dependencies. The total count of vulnerabilities was 83, including 21 critical ones.

The `socket.io` package was upgraded from version `^2.0.4` to `^4.7.2`. This upgrade, however, has introduced a bug affecting the connection to the Front End, which uses `socket.io-client` version `2.x.x`. This issue has been documented in Issue #1.

Further action is required to ensure compatibility between the Front End and the updated `socket.io` version.