fix(client): fix signup failure bug (#25)
This fixes part of issue #14. When a user signed up the client used to always display failure even if it succeeded in the BE and created a user. It turns out that this was an issue in the comparison response.status = 200, as instead of examining the response code, we were examining the response's field status which contained either "success" or "failure", this was because obeserve: response was missing which allows us to examine status codes in the response.