feat(cicd): only run CICD on push (#50)
This repo is about to become public and we must prevent unauthorized deployment of code. This is achived by only deploying on push, as we can review user's pull requests to develop, then, they will only be deployed if approved, and it this goes well, dev will be merge into main and deployed.